Contextualize the blobstore API
Pass in a context as the first arugment to Get, Put, Contains, and
Delete. Nothing checks the status of the context, but I may add that in
a future pull request.

The S3 blobstore correct uses the `WithContext` AWS SDK methods.

The README has been updated with new contextualized API.